requirements.txt
numpy==1.26.4
openai-whisper==20231117
openai==1.30.1
streamlit
tiktoken==0.7.0

